Change location of queries and add function to subscribe and get subcribers from channels etc
