+++ services/ChannelsService.js
@@ -1,72 +1,126 @@
-const channelQueries = require('../database/queries/ChannelQueries');
+const database = require('../database/db');
+const db = database.initDatabase();
 
-const createChannel = async (req, res) => {
-  const { name, description, userId } = req.body;
-
-  try {
-    await channelQueries.create_Channel(name, description, userId);
-    res.send('Channel successfully created');
-  } catch (error) {
-    console.log(error);
-    res.status(404).send('Internal server error');
-    res.status(404).send('Internal server error');
-  }
+const createChannel = async (name, description, userId) => {
+  return new Promise((resolve, reject) => {
+    db.run(
+      `INSERT INTO channels (Name, Description, Owner_id) VALUES (?,?,?)`,
+      [name, description, userId],
+      function (error) {
+        if (error) {
+          console.log(error);
+          reject(error);
+        } else {
+          console.log(`Created channel ${name} successfully`);
+          resolve();
+        }
+      }
+    );
+  });
 };
 
-const getAllChannels = async (req, res) => {
-  try {
-    const channels = await channelQueries.getAll_Channels();
-    console.log(channels);
-    res.json(channels);
-  } catch (error) {
-    console.log(error);
-    res.status(404).send('Internal server error');
-    res.status(404).send('Internal server error');
-  }
+const getAllChannels = async () => {
+  return new Promise((resolve, reject) => {
+    db.all(`SELECT * FROM channels`, [], (error, channels) => {
+      if (error) {
+        console.error(error);
+        reject(error);
+      } else {
+        resolve(channels);
+      }
+    });
+  });
 };
 
-const updateChannel = async (req, res) => {
-  const { name, description, channelId } = req.body;
-
-  try {
-    await channelQueries.update_Channel(name, description, channelId);
-    res.send('Channel successfully updated');
-  } catch (error) {
-    console.log(error);
-    res.status(404).send('Internal server error');
-    res.status(404).send('Internal server error');
-  }
+const updateChannel = async (name, description, channelId) => {
+  return new Promise((resolve, reject) => {
+    db.run(
+      `UPDATE channels SET Name = ?, Description = ? WHERE Id = ?`,
+      [name, description, channelId],
+      function (error) {
+        if (error) {
+          console.error(error);
+          reject(error);
+        } else {
+          console.log(`Channel successfully updated`);
+          resolve();
+        }
+      }
+    );
+  });
 };
 
-const deleteChannel = async (req, res) => {
-  const { channelId } = req.body;
-
-  try {
-    await channelQueries.delete_Channel(channelId);
-    res.send('Channel successfully deleted');
-  } catch (error) {
-    console.log(error);
-    res.status(404).send('Internal server error');
-  }
+const deleteChannel = async (channelId) => {
+  return new Promise((resolve, reject) => {
+    db.run(`DELETE FROM channels WHERE Id = ?`, [channelId], function (error) {
+      if (error) {
+        console.error(error);
+        reject(error);
+      } else {
+        console.log(`Channel deleted`);
+        resolve();
+      }
+    });
+  });
 };
 
-const getAllChannelMessages = async (req, res) => {
-  const { id: channelId } = req.params;
+const getAllChannelMessages = async (channelId) => {
+  return new Promise((resolve, reject) => {
+    db.all(
+      `SELECT messages.* FROM messages JOIN messagesChannels ON messages.Id = messagesChannels.Message_Id WHERE messagesChannels.Channel_Id = ?`,
+      [channelId],
+      function (error, rows) {
+        if (error) {
+          console.error(error);
+          reject(error);
+        } else {
+          resolve(rows);
+        }
+      }
+    );
+  });
+};
 
-  try {
-    const messages = await channelQueries.getAll_ChannelMessages(channelId);
-    res.json(messages);
-  } catch (error) {
-    console.log(error);
-    res.status(404).send('Internal server error');
-    res.status(404).send('Internal server error');
-  }
+const getAllChannelUsers = async (channelId) => {
+  return new Promise((resolve, reject) => {
+    db.all(
+      `SELECT users. * FROM users JOIN usersChannels ON users.Id = usersChannels.User_Id WHERE usersChannels.Channel_Id = ?`,
+      [channelId],
+      function (error, rows) {
+        if (error) {
+          console.error(error);
+          reject(error);
+        } else {
+          resolve(rows);
+        }
+      }
+    );
+  });
+};
+
+const subscribeToChannel = async (userId, channelId) => {
+  return new Promise((resolve, reject) => {
+    db.run(
+      `INSERT INTO usersChannels (User_Id, Channel_Id) VALUES (?,?)`,
+      [userId, channelId],
+      function (error) {
+        if (error) {
+          console.error(error);
+          reject(error);
+        } else {
+          resolve();
+        }
+      }
+    );
+  });
 };
 
 module.exports = {
   createChannel,
-  deleteChannel,
-  updateChannel,
   getAllChannels,
+  updateChannel,
+  deleteChannel,
   getAllChannelMessages,
+  getAllChannelUsers,
+  subscribeToChannel,
 };
